requirements-old.txt
django
djangorestframework
django-cors-headers
ConfigParser
mysqlclient